delete edit, add forms added edit not working yet

+++ ngPetConnect/src/app/component/mypets/mypets.component.ts
@@ -1,12 +1,10 @@
 import { AuthService } from './../../services/auth.service';
-import { User } from './../../models/user';
 import { PetService } from './../../services/pet.service';
 import { Component, OnInit } from '@angular/core';
 import { HomeComponent } from '../home/home.component';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, NgModel } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { Pet } from '../../models/pet';
-import { UserService } from '../../services/user.service';
 
 
 
@@ -24,6 +22,7 @@ export class MypetsComponent implements OnInit {
   selected: Pet | null = null;
   editPet: Pet | null = null;
   pet: Pet[] = [];
+  addPetClicked: boolean = false;
 
   constructor(private petService: PetService, private authService: AuthService) {}
 
@@ -47,7 +46,8 @@ export class MypetsComponent implements OnInit {
     }
 }
 
-addInventory() {
+addPets() {
+  if(this.authService.checkLogin()) {
   this.petService.create(this.newPet).subscribe(
     (createdPet) => {
       this.pet.push(createdPet);
@@ -58,20 +58,25 @@ addInventory() {
       console.error('Error creating pet:', error);
     }
   );
+  }
 }
 
-deleteInventory(id: number) {
-  this.petService.destroy(id).subscribe(
-    () => {
-      this.pet = this.pet.filter((pet) => pet.id !== id);
+deletePet(id: number) {
+  if(this.authService.checkLogin()) {
+    this.petService.destroy(id).subscribe(
+      () => {
+        this.pet = this.pet.filter((pet) => pet.id !== id);
+        console.log('Delete of pet id: ' + id + ' successful')
     },
     (error) => {
       console.error('Error deleting pet:', error);
     }
   );
+  }
 }
 
-updateInventory() {
+updatePets() {
+  if(this.authService.checkLogin()) {
   if (this.editPet) {
     this.petService.update(this.editPet).subscribe(
       () => {
@@ -85,6 +90,7 @@ updateInventory() {
     );
   }
 }
+}
 
 displayPets(item: Pet): void {
   this.selected = item;
@@ -94,8 +100,13 @@ displayTable(): void {
   this.selected = null;
 }
 
-setEditPets() {
+setEditPets(pet: Pet) {
   this.isEditing = true;
-  this.editPet = Object.assign({}, this.selected);
+  // this.editPet = Object.assign({}, this.selected);
+  this.editPet = { ...pet};
 }
+  toggleAddPet() {
+    this.addPetClicked = !this.addPetClicked;
+  }
+
 }
